Comments on raster vs vector
The difference between raster and vector is that raster graphic store information for each pixels, that is why when we zoom in the picture looks very pixelated. While vector graphics store multiple mathematics curve telling what point connect to what point so that when we zoom in it can do more calculation to increase the resolution of the picture so it doesnt look pixelated.